refactor: extract external api
