+++ src/main/java/org/hoongoin/interviewbank/account/domain/AccountCommandService.java
@@ -7,32 +7,15 @@ import org.hoongoin.interviewbank.account.infrastructure.entity.AccountEntity;
 import org.hoongoin.interviewbank.account.application.entity.AccountType;
 import org.hoongoin.interviewbank.account.infrastructure.repository.AccountRepository;
 import org.hoongoin.interviewbank.account.application.entity.Account;
-import org.hoongoin.interviewbank.exception.IbBadRequestException;
 import org.hoongoin.interviewbank.exception.IbEntityExistsException;
 import org.hoongoin.interviewbank.exception.IbEntityNotFoundException;
-import org.hoongoin.interviewbank.exception.IbValidationException;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
 
-import javax.imageio.ImageIO;
 import javax.transaction.Transactional;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.net.URI;
 import java.util.Optional;
-import java.util.UUID;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
 
 @Service
 @RequiredArgsConstructor
@@ -40,12 +23,6 @@ public class AccountCommandService {
 
 	private final AccountRepository accountRepository;
 	private final AccountMapper accountMapper;
-	private final AmazonS3 amazonS3;
-
-	@Value("${aws.s3.bucket}")
-	private String bucket;
-	@Value("${aws.region.static}")
-	private String region;
 
 	public Account insertAccount(Account account) {
 		Boolean emailExists = accountRepository.existsByEmailAndAccountType(account.getEmail(), AccountType.EMAIL);
@@ -105,92 +82,12 @@ public class AccountCommandService {
 		return accountMapper.accountEntityToAccount(accountEntity);
 	}
 
-	public Account updateImageUrl(long requestedAccountId, MultipartFile multipartFile) {
+	public Account updateImageUrl(long requestedAccountId, String uploadedUrl) {
 		AccountEntity accountEntity = accountRepository.findById(requestedAccountId)
 			.orElseThrow(() -> new IbEntityNotFoundException("Account"));
 
-		checkImageUrlOfAccountEntity(accountEntity);
-
-		byte[] fileBytes = resizeImageSizeOfFile(multipartFile);
-
-		String filename = UUID.randomUUID().toString();
-
-		saveFileToS3(multipartFile, fileBytes, filename);
-
-		String extractedS3Url = extractS3Url(filename);
-
-		accountEntity.uploadImageUrl(extractedS3Url);
+		accountEntity.uploadImageUrl(uploadedUrl);
 
 		return accountMapper.accountEntityToAccount(accountEntity);
-	}
-
-	private byte[] resizeImageSizeOfFile(MultipartFile multipartFile) {
-		BufferedImage originalImage;
-
-		try {
-			originalImage = ImageIO.read(multipartFile.getInputStream());
-		} catch (IOException e) {
-			throw new IbBadRequestException(e.getMessage());
-		}
-
-		int resizedWidth = resize(originalImage.getWidth());
-		int resizedHeight = resize(originalImage.getHeight());
-
-		BufferedImage resizedImage = new BufferedImage(resizedWidth, resizedHeight, originalImage.getType());
-		Graphics2D graphics = resizedImage.createGraphics();
-		graphics.drawImage(originalImage, 0, 0, resizedWidth, resizedHeight, null);
-		graphics.dispose();
-
-		return imageAndContentTypeToBytes(resizedImage, multipartFile.getContentType());
-	}
-
-	private int resize(int size) {
-		return Math.min(size, 400);
-	}
-
-	private byte[] imageAndContentTypeToBytes(BufferedImage image, String contentType) {
-		try {
-			ByteArrayOutputStream bos = new ByteArrayOutputStream();
-			ImageIO.write(image, getFormatName(contentType), bos);
-			return bos.toByteArray();
-		} catch (IOException e) {
-			throw new IbBadRequestException(e.getMessage());
-		}
-	}
-
-	private String getFormatName(String contentType) {
-		return switch (contentType) {
-			case "image/jpeg" -> "jpeg";
-			case "image/png" -> "png";
-			case "image/gif" -> "gif";
-			default -> throw new IllegalArgumentException("Unsupported image type: " + contentType);
-		};
-	}
-
-	private void saveFileToS3(MultipartFile multipartFile, byte[] fileBytes,
-		String filename) {
-		ObjectMetadata metadata = new ObjectMetadata();
-		metadata.setContentLength(fileBytes.length);
-		metadata.setContentType(multipartFile.getContentType());
-		amazonS3.putObject(new PutObjectRequest(bucket, filename, new ByteArrayInputStream(fileBytes), metadata));
-	}
-
-	private String extractS3Url(String filename) {
-		return "https://" + bucket + ".s3." + region + ".amazonaws.com/" + filename;
-	}
-
-	private void checkImageUrlOfAccountEntity(AccountEntity accountEntity) {
-		if (accountEntity.getImageUrl() != null) {
-			amazonS3.deleteObject(new DeleteObjectRequest(bucket, getObjectKeyFromS3Url(accountEntity.getImageUrl())));
-		}
-	}
-
-	public String getObjectKeyFromS3Url(String url) {
-		try {
-			URI uri = new URI(url);
-			return uri.getPath().substring(1);
-		} catch (Exception e) {
-			throw new IbValidationException("invalid uri format");
-		}
 	}
 }
